requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
coloredlogs==15.0.1
fastapi==0.110.3
fasteners==0.19
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
humanfriendly==10.0
idna==3.7
magic-filter==1.0.12
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
property-manager==3.0
psycopg2==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dotenv==1.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
Telethon==1.35.0
typing_extensions==4.11.0
uvicorn==0.29.0
verboselogs==1.7
yarl==1.9.4
